# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutRedirect": {
        "additionalProperties": false,
        "properties": {
          "checkout_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Checkout Url",
            "type": "string"
          },
          "provider": {
            "const": "creem",
            "title": "Provider",
            "type": "string"
          },
          "purchase_id": {
            "format": "uuid",
            "title": "Purchase Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "READY",
              "CONFIRMED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "purchase_id",
          "provider",
          "status",
          "checkout_url"
        ],
        "title": "CheckoutRedirect",
        "type": "object"
      },
...
      "CreditPackCheckoutRequest": {
        "additionalProperties": false,
        "properties": {
          "product_code": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Product Code",
            "type": "string"
          },
          "return_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Url"
          }
        },
        "required": [
          "product_code"
        ],
        "title": "CreditPackCheckoutRequest",
        "type": "object"
      },
      "CreditPackStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "accounting_balance": {
            "title": "Accounting Balance",
            "type": "integer"
          },
          "captured_minor_units": {
            "minimum": 0.0,
            "title": "Captured Minor Units",
            "type": "integer"
          },
          "checkout_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Url"
          },
          "credits_granted": {
            "minimum": 0.0,
            "title": "Credits Granted",
            "type": "integer"
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "disputed_minor_units": {
            "minimum": 0.0,
            "title": "Disputed Minor Units",
            "type": "integer"
          },
          "product_code": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Product Code",
            "type": "string"
          },
          "provider": {
            "const": "creem",
            "title": "Provider",
            "type": "string"
          },
          "purchase_id": {
            "format": "uuid",
            "title": "Purchase Id",
            "type": "string"
          },
          "refunded_minor_units": {
            "minimum": 0.0,
            "title": "Refunded Minor Units",
            "type": "integer"
          },
          "spendable_balance": {
            "minimum": 0.0,
            "title": "Spendable Balance",
            "type": "integer"
          },
          "state": {
            "enum": [
              "PENDING",
              "PAID",
              "PARTIAL_RECONCILIATION_REQUIRED",
              "DISPUTED",
              "REVERSED",
              "UNKNOWN"
            ],
            "title": "State",
            "type": "string"
          },
          "tax_minor_units": {
            "minimum": 0.0,
            "title": "Tax Minor Units",
            "type": "integer"
          }
        },
        "required": [
          "purchase_id",
          "provider",
          "product_code",
          "state",
          "captured_minor_units",
          "tax_minor_units",
          "refunded_minor_units",
          "disputed_minor_units",
          "currency",
          "credits_granted",
          "accounting_balance",
          "spendable_balance"
        ],
        "title": "CreditPackStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API for AI Wedding Photo generation using InstantID",
    "title": "AI Wedding Photo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/payments/checkout": {
      "post": {
        "operationId": "create_checkout_api_v1_payments_checkout_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditPackCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutRedirect"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Checkout",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/status/{purchase_id}": {
      "get": {
        "operationId": "get_payment_status_api_v1_payments_status__purchase_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "title": "Purchase Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "checkout_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Checkout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditPackStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment Status",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/api/v1/payments/{purchase_id}/refund": {
      "post": {
        "operationId": "initiate_refund_api_v1_payments__purchase_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Initiate Refund Api V1 Payments  Purchase Id  Refund Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Initiate Refund",
        "tags": [
          "payments"
        ]
      }
    },
...
  }
}